\name{getSparse}
\alias{getSparse}
\title{Getter of the intermediate/output DelayedArray object in DelayedTensor}
\usage{
getSparse()
}
\value{
TRUE or FALSE (Default: TRUE)
}
\description{
Whether the intermediate and output DelayedArray used in DelayedTensor
is used as sparse tensor or not.
}
\examples{
getSparse()
}
