\name{setVerbose}
\alias{setVerbose}
\title{Setter to set the verbose mode of DelayedTensor}
\usage{
setVerbose(as.verbose=FALSE)
}
\arguments{
\item{as.verbose}{TRUE or FALSE (Default: FALSE)}
}
\value{Nothing}
\description{
Set the verbose message to monitor the block-processing procedure.
}
\examples{
setVerbose(TRUE)
setVerbose(FALSE)
}
